src/TwoFABot/requirements.txt
pgpy==0.6.0
python-telegram-bot==13.13
